rename: 서비스명 변경 (recommendation -> ai)
