+++ ai/.gitattributes
