+++ ai/src/main/java/com/klp/ai/AiApplication.java
@@ -1,0 +1,17 @@
+package com.klp.ai;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@SpringBootApplication
+@EnableFeignClients
+@EnableAsync
+public class AiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AiApplication.class, args);
+    }
+
+}
