--- recommendation/src/main/java/com/klp/recommendation/RecommendationApplication.java
@@ -1,13 +1,0 @@
-package com.klp.recommendation;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RecommendationApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(RecommendationApplication.class, args);
-    }
-
-}
